refactor!: remove prefix from open_feature_* modules
