+++ open_feature/client.py
@@ -2,7 +2,7 @@ import logging
 import typing
 from dataclasses import dataclass
 
-from open_feature import open_feature_api as api
+from open_feature import api
 from open_feature.evaluation_context import EvaluationContext
 from open_feature.exception import (
     ErrorCode,
