+++ readme.md
@@ -66,11 +66,11 @@ pip install requirements.txt
 In order to use the sdk there is some minor configuration. Follow the script below:
 
 ```python
-from open_feature import open_feature_api
+from open_feature import api
 from open_feature.provider.no_op_provider import NoOpProvider
 
-open_feature_api.set_provider(NoOpProvider())
-open_feature_client = open_feature_api.get_client()
+api.set_provider(NoOpProvider())
+open_feature_client = api.get_client()
 ```
 
 ### Basics:
@@ -91,7 +91,7 @@ You can also bind a provider to a specific client by name instead of setting tha
 
 ```python
 
-open_feature_api.set_provider(NoOpProvider())
+api.set_provider(NoOpProvider())
 ```
 
 Each provider class may have further setup required i.e. secret keys, environment variables etc
@@ -103,7 +103,7 @@ In OpenFeature, we refer to this as [`targeting`](https://openfeature.dev/specif
 If the flag system you're using supports targeting, you can provide the input data using the `EvaluationContext`.
 
 ```python
-from open_feature.open_feature_api import (
+from open_feature.api import (
     get_client,
     get_provider,
     set_provider
@@ -150,7 +150,7 @@ class MyHook(Hook):
 
 
 # set global hooks at the API-level
-from open_feature.open_feature_api import add_hooks
+from open_feature.api import add_hooks
 add_hooks([MyHook()])
 
 # or configure them in the client
