+++ tests/conftest.py
@@ -1,6 +1,6 @@
 import pytest
 
-from open_feature import open_feature_api as api
+from open_feature import api
 from open_feature.provider.no_op_provider import NoOpProvider
